Delete all LLVM interop from ILC
