+++ src/coreclr/tools/aot/ILCompiler.LLVM/ILCompiler.LLVM.csproj
@@ -21,9 +21,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Include="CodeGen\LLVMCodegenCompilation.CodeGen.cs" />
-    <Compile Include="CodeGen\LLVMObjectWriter.cs" />
     <Compile Include="CodeGen\WasmObjectWriter.cs" />
+    <Compile Include="CodeGen\WasmObjectWriter.EmitObject.cs" />
     <Compile Include="CodeGen\WasmNative.cs" />
     <Compile Include="Compiler\DependencyAnalysis\ExternMethodAccessorNode.cs" />
     <Compile Include="Compiler\DependencyAnalysis\LLVMCodegenNodeFactory.cs" />
